实现定时删除交易hash和区块； 实现统计每日交易数据； 实现每日每日交易数据查询（不够七天补齐到七天）

+++ src/main/java/com/webank/fabric/node/manager/common/pojo/transaction/TransactionDO.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class TransactionDO {
-    private Integer transNumber;
+    private BigInteger transNumber;
     private String txId;
     private BigInteger blockNumber;
     private int actionCount;
